chore: refactor network/score files
