+++ packages/beacon-node/src/network/peers/score/index.ts
@@ -1,0 +1,4 @@
+export * from "./store.js";
+export * from "./interface.js";
+export * from "./score.js";
+export * from "./utils.js";
