+++ packages/beacon-node/src/network/reqresp/ReqRespBeaconNode.ts
@@ -19,7 +19,7 @@ import {NetworkCoreMetrics} from "../core/metrics.js";
 import {INetworkEventBus, NetworkEvent} from "../events.js";
 import {MetadataController} from "../metadata.js";
 import {PeersData} from "../peers/peersData.js";
-import {IPeerRpcScoreStore, PeerAction} from "../peers/score.js";
+import {IPeerRpcScoreStore, PeerAction} from "../peers/score/index.js";
 import {ReqRespHandlers} from "./handlers/index.js";
 import {IReqRespBeaconNode} from "./interface.js";
 import {onOutgoingReqRespError} from "./score.js";
