+++ packages/beacon-node/src/network/reqresp/score.ts
@@ -1,5 +1,5 @@
 import {RequestError, RequestErrorCode} from "@lodestar/reqresp";
-import {PeerAction} from "../peers/score.js";
+import {PeerAction} from "../peers/score/index.js";
 import {ReqRespMethod} from "./types.js";
 
 /**
